refactor: consolidate agent code under internal
